Change role for modernisation platform to readonly
The view only role was completely unusable, the readonly role should
give us the visibility we need to manage the platform account.  If we
ever need admin permissions outside of the pipelines we can use the
account superadmin users.  Hopefully if this role is more usable there
will be less logging in as superadmins.