Add Modernisation Platform Engineer role
Currently the Modernisation Platform generally use an Admin role to
investigate issues via the console for both core and member accounts.  This is not best practice, this is
an attempt to try out using a read only role with expanded permissions
where needed, eg to raise a support ticket for an account.

If we can start using an account like this for more general trouble
shooting it reduces the risk of accidental or intentional clickops
changes.

Engineers would still have the option of assuming a Admin or superadmin
role if needed.