Add Cloudformation stack for Oracle licensing
Following this documentation

https://aws.amazon.com/blogs/mt/centrally-track-oracle-database-licenses-in-aws-organizations-using-aws-license-manager-and-aws-systems-manager/

Need to question the need for an InstanceIAMRole, assuming this is the
role attached to an instance profile, there can only be one role per
profile.  So all the instances with Oracle on them would need to have
the same instance profile role name which seems odd.  Also this
parameter isn't mentioned in the blog post.