Add ownership controls to S3 buckets
Updates to to S3 buckets have ACLs disabled by default.  Terraform defaults the bucket ownership controls to BucketOwnerEnforced now which breaks when trying to create buckets with ACLS. Defining this to BucketOwnerEnforced fixes this.  We should look at disabling the ACLs completely at some point, but changing this in the module would impact existing buckets and needs to be fully understood before changing.