Ignore enable_default_standards changes
Once security hub has been enabled there is no way to amend this value, when trying to import security hub which was already enabled it was set to false.  This requried the resource to be destroyed which isn't possible unless you remove the delegation to this account which affects all accounts.