Remove logic and role assumption in org-sec providers
When running this tf on a pipeline it uses the github actions plan or apply role, and uses the permissions associated with that role.

When running this terraform locally, it assumes the OrganizationAccountAccessRole in the organisation-security account, as we normally run this with the root account credentials.  This was because historically, we didn't have organisation-security credentials via SSO, but we do now.

This change removes the logic, so the terraform will run with whatever permissions the credentials that run it have, this doesn't change anything for running on the pipeline, but for local terraform runs you will now need to run it with the organisation-security credentials.

This makes more sense, plus we don't have to use root credentials unecessarily any more.