Enable Runtime monitoring for new accounts
We do not want to do this for all accounts currently, this way we have this configuration in code, but we can choose which accounts to add to this.

Also with the agent configurations, we do not want to automatically add the agents we need to add them and test their impact.